+++ src/test/java/com/jnape/palatable/lambda/monoid/builtin/AndTest.java
@@ -2,6 +2,9 @@ package com.jnape.palatable.lambda.monoid.builtin;
 
 import org.junit.Test;
 
+import static com.jnape.palatable.lambda.functions.builtin.fn1.Id.id;
+import static com.jnape.palatable.lambda.functions.builtin.fn1.Repeat.repeat;
+import static com.jnape.palatable.lambda.functions.builtin.fn2.Cons.cons;
 import static com.jnape.palatable.lambda.monoid.builtin.And.and;
 import static org.junit.Assert.assertEquals;
 
@@ -19,5 +22,16 @@ public class AndTest {
         assertEquals(false, and.apply(false, true));
         assertEquals(false, and.apply(true, false));
         assertEquals(false, and.apply(false, false));
+    }
+
+    @Test(timeout = 500)
+    public void shortCircuiting() {
+        Iterable<Boolean> bools = cons(false, repeat(true));
+        And and = and();
+
+        assertEquals(false, and.foldLeft(false, bools));
+        assertEquals(false, and.foldLeft(true, bools));
+        assertEquals(false, and.reduceLeft(bools));
+        assertEquals(false, and.foldMap(id(), bools));
     }
 }